@@ -2,6 +2,5 @@
 # Controller for the homepage. Since there is not much logic on the homepage this controller is
 # rather empty.
 class HomeController < ApplicationController
-  def index
-  end
+  def index; end
 end
