@@ -7,7 +7,7 @@ class TeamsControllerTest < ActionController::TestCase
   end
 
   test 'unauthenticated users should not be able to access team show page' do
-    get :show, id: teams(:team_one)
+    get :show, params: { id: teams(:team_one) }
     assert_redirected_to root_path
   end
 
@@ -27,14 +27,14 @@ class TeamsControllerTest < ActionController::TestCase
   test 'authenticated users with a team can view their team' do
     user = users(:user_one)
     sign_in user
-    get :show, id: user.team
+    get :show, params: { id: user.team }
     assert_response :success
   end
 
   test 'authenticated users with a team cannot view other teams' do
     user = users(:user_one)
     sign_in user
-    get :show, id: teams(:team_two)
+    get :show, params: { id: teams(:team_two) }
     assert_redirected_to team_path(user.team)
     assert_equal flash[:alert], I18n.t('teams.invalid_permissions')
   end
@@ -42,14 +42,14 @@ class TeamsControllerTest < ActionController::TestCase
   test 'authenticated users without a team cannot view other teams' do
     user = users(:user_two)
     sign_in user
-    get :show, id: teams(:team_two)
+    get :show, params: { id: teams(:team_two) }
     assert_redirected_to @controller.user_root_path
   end
 
   test 'a team cannot be created with the same name as another team' do
     sign_in users(:user_two)
     assert_no_difference 'Team.count', 'A Team should not be created' do
-      post :create, team: { team_name: 'team_one', affiliation: 'school1' }
+      post :create, params: { team: { team_name: 'team_one', affiliation: 'school1' } }
     end
     assert_template :new
   end
@@ -57,7 +57,7 @@ class TeamsControllerTest < ActionController::TestCase
   test 'a team cannot be created with the same name as another team in any case' do
     sign_in users(:user_two)
     assert_no_difference 'Team.count', 'A Team should not be created' do
-      post :create, team: { team_name: 'TeAm_OnE', affiliation: 'school1' }
+      post :create, params: { team: { team_name: 'TeAm_OnE', affiliation: 'school1' } }
     end
     assert_template :new
   end
@@ -66,7 +66,7 @@ class TeamsControllerTest < ActionController::TestCase
     user = users(:user_two)
     sign_in user
     assert_difference 'Team.count' do
-      post :create, team: { team_name: 'another_team', affiliation: 'school1' }
+      post :create, params: { team: { team_name: 'another_team', affiliation: 'school1' } }
     end
     user.reload
     assert_redirected_to team_path(user.team)
@@ -76,14 +76,14 @@ class TeamsControllerTest < ActionController::TestCase
   test 'a team captain of a full team sees a message informing them that their team is full' do
     user = users(:full_team_user_one)
     sign_in user
-    get :show, id: user.team
+    get :show, params: { id: user.team }
     assert_equal flash[:notice], I18n.t('teams.full_team')
   end
 
   test 'invite a team member' do
     user = users(:user_one)
     sign_in user
-    patch :update, team: { team_name: 'team_one', affiliation: 'school1' }, id: user.team
+    patch :update, params: { team: { team_name: 'team_one', affiliation: 'school1' }, id: user.team }
     assert_redirected_to team_path(users(:user_one).team)
     assert I18n.t('invites.invite_successful'), flash[:notice]
   end
