@@ -9,14 +9,14 @@ class UserRequestsControllerTest < ActionController::TestCase
 
   test 'user cannot create a request with one already pending for same team' do
     sign_in users(:user_two)
-    post :create, team_id: teams(:team_one)
+    post :create, params: { team_id: teams(:team_one) }
     assert_redirected_to join_team_users_path
     assert_match I18n.t('requests.already_pending'), flash[:alert]
   end
 
   test 'user can create request when they do not have any pending requests for that team and are not on a team' do
     sign_in users(:user_three)
-    post :create, team_id: teams(:team_one)
+    post :create, params: { team_id: teams(:team_one) }
     assert_redirected_to join_team_users_path
     assert_equal I18n.t('requests.sent_successful'), flash[:notice]
   end
@@ -27,7 +27,7 @@ class UserRequestsControllerTest < ActionController::TestCase
   test 'accept request' do
     user_request = user_requests(:request_one)
     sign_in users(:user_one)
-    get :accept, team_id: user_request.team, id: user_request
+    get :accept, params: { team_id: user_request.team, id: user_request }
     assert_redirected_to @controller.user_root_path
     assert_equal I18n.t('requests.accepted_successful'), flash[:notice]
   end
@@ -35,16 +35,34 @@ class UserRequestsControllerTest < ActionController::TestCase
   test 'accept request but user already on a team' do
     user_request = user_requests(:request_three)
     sign_in users(:user_one)
-    get :accept, team_id: user_request.team, id: user_request
+    get :accept, params: { team_id: user_request.team, id: user_request }
     assert_redirected_to @request.env["HTTP_REFERER"]
+    assert_equal I18n.t('requests.accepted_another'), flash[:alert]
+  end
+
+  test 'accept request but user already on a team with no referer' do
+    user_request = user_requests(:request_three)
+    sign_in users(:user_one)
+    @request.env["HTTP_REFERER"] = nil
+    get :accept, params: { team_id: user_request.team, id: user_request }
+    assert_redirected_to @controller.user_root_path
     assert_equal I18n.t('requests.accepted_another'), flash[:alert]
   end
 
   test 'accept request but team now has too many members' do
     user_request = user_requests(:request_two)
     sign_in users(:full_team_user_one)
-    get :accept, team_id: user_request.team, id: user_request
+    get :accept, params: { team_id: user_request.team, id: user_request }
     assert_redirected_to @request.env["HTTP_REFERER"]
+    assert_equal I18n.t('requests.too_many_players'), flash[:alert]
+  end
+
+  test 'accept request but team now has too many members with no referer' do
+    user_request = user_requests(:request_two)
+    sign_in users(:full_team_user_one)
+    @request.env["HTTP_REFERER"] = nil
+    get :accept, params: { team_id: user_request.team, id: user_request }
+    assert_redirected_to @controller.user_root_path
     assert_equal I18n.t('requests.too_many_players'), flash[:alert]
   end
 
@@ -52,23 +70,41 @@ class UserRequestsControllerTest < ActionController::TestCase
     user_request = user_requests(:request_four)
     sign_in users(:user_six)
     assert_raise ActiveRecord::RecordNotFound do
-      get :accept, team_id: user_request.team, id: user_request
+      get :accept, params: { team_id: user_request.team, id: user_request }
     end
   end
 
   test 'user destroys own request' do
     user_request = user_requests(:request_one)
     sign_in users(:user_two)
-    delete :destroy, team_id: user_request.team, id: user_request
+    delete :destroy, params: { team_id: user_request.team, id: user_request }
     assert_redirected_to @request.env["HTTP_REFERER"]
+    assert_equal I18n.t('requests.rejected_successful'), flash[:notice]
+  end
+
+  test 'user destroys own request no referer' do
+    user_request = user_requests(:request_one)
+    sign_in users(:user_two)
+    @request.env["HTTP_REFERER"] = nil
+    delete :destroy, params: { team_id: user_request.team, id: user_request }
+    assert_redirected_to @controller.user_root_path
     assert_equal I18n.t('requests.rejected_successful'), flash[:notice]
   end
 
   test 'captain destroys user request' do
     user_request = user_requests(:request_one)
     sign_in users(:user_one)
-    delete :destroy, team_id: user_request.team, id: user_request
+    delete :destroy, params: { team_id: user_request.team, id: user_request }
     assert_redirected_to @request.env["HTTP_REFERER"]
+    assert_equal I18n.t('requests.rejected_successful'), flash[:notice]
+  end
+
+  test 'captain destroys user request no referer' do
+    user_request = user_requests(:request_one)
+    sign_in users(:user_one)
+    @request.env["HTTP_REFERER"] = nil
+    delete :destroy, params: { team_id: user_request.team, id: user_request }
+    assert_redirected_to @controller.user_root_path
     assert_equal I18n.t('requests.rejected_successful'), flash[:notice]
   end
 
@@ -76,7 +112,7 @@ class UserRequestsControllerTest < ActionController::TestCase
     user_request = user_requests(:request_four)
     sign_in users(:user_five)
     assert_raise ActiveRecord::RecordNotFound do
-      delete :destroy, team_id: user_request.team, id: user_request
+      delete :destroy, params: { team_id: user_request.team, id: user_request }
     end
   end
 end
